LPS-123495 Inline the only usage
