+++ modules/apps/portal-cache/portal-cache-multiple/src/main/java/com/liferay/portal/cache/multiple/internal/portal/profile/ModulePortalProfile.java
@@ -16,7 +16,6 @@ package com.liferay.portal.cache.multiple.internal.portal.profile;
 
 import com.liferay.portal.cache.PortalCacheReplicatorFactory;
 import com.liferay.portal.cache.multiple.internal.ClusterLinkPortalCacheReplicatorFactory;
-import com.liferay.portal.cache.multiple.internal.PortalCacheManagerUtil;
 import com.liferay.portal.cache.multiple.internal.cluster.link.ClusterLinkPortalCacheClusterChannelFactory;
 import com.liferay.portal.cache.multiple.internal.cluster.link.PortalCacheClusterLink;
 import com.liferay.portal.cache.multiple.internal.cluster.link.messaging.ClusterLinkMessagingConfigurator;
@@ -73,8 +72,7 @@ public class ModulePortalProfile extends BaseDSModulePortalProfile {
 			ClusterLinkPortalCacheClusterChannelFactory.class.getName(),
 			ClusterLinkPortalCacheClusterListener.class.getName(),
 			ClusterLinkPortalCacheReplicatorFactory.class.getName(),
-			PortalCacheClusterLink.class.getName(),
-			PortalCacheManagerUtil.class.getName());
+			PortalCacheClusterLink.class.getName());
 	}
 
 	@Reference
